move repl into a separate folder
